Move makeWrapper to global test utils.
